refactor(frontend): split playtest character runtime
